+++ source/ReadAlign.cpp
@@ -90,6 +90,7 @@ ReadAlign::ReadAlign (Parameters* Pin, const Genome &genomeIn, Transcriptome *Tr
         readNameMates[ii]=new char [DEF_readNameLengthMax];
     };
 
+    readNameExtra.resize(P->readNmates);
 
     outBAMoneAlignNbytes = new uint [P->readNmates+2]; //extra piece for chimeric reads
     outBAMoneAlign = new char* [P->readNmates+2]; //extra piece for chimeric reads