+++ source/ReadAlign_alignBAM.cpp
@@ -55,6 +55,12 @@ int bamAttrArrayWrite(int32 attr, const char* tagName, char* attrArray ) {
     *( (int32*) (attrArray+3))=attr;
     return 3+sizeof(int32);
 };
+int bamAttrArrayWrite(float attr, const char* tagName, char* attrArray ) {
+    attrArray[0]=tagName[0];attrArray[1]=tagName[1];
+    attrArray[2]='f';
+    *( (float*) (attrArray+3))=attr;
+    return 3+sizeof(int32);
+};
 int bamAttrArrayWrite(char attr, const char* tagName, char* attrArray ) {
     attrArray[0]=tagName[0];attrArray[1]=tagName[1];
     attrArray[2]='A';
@@ -82,6 +88,49 @@ int bamAttrArrayWrite(vector<int32> &attr, const char* tagName, char* attrArray 
     *( (int32*) (attrArray+4))=attr.size();
     memcpy(attrArray+4+sizeof(int32),attr.data(),sizeof(int32)*attr.size());//copy array data
     return 4+sizeof(int32)+sizeof(int32)*attr.size();
+};
+
+int bamAttrArrayWriteSAMtags(string &attrStr, char *attrArray) {//write bam record into attrArray for string attribute attString
+    size_t pos1=0, pos2=0;
+    int nattr=0;
+    do {//cycle over multiple tags separated by tab
+        pos2 = attrStr.find('\t',pos1);
+        string attr1 = attrStr.substr(pos1, pos2-pos1);
+        pos1=pos2+1;
+        
+        if (attr1.empty())
+            continue; //extra tab at the beginning, or consecutive tabs
+        
+        switch (attr1.at(3)) {
+            case 'i':
+            {
+                int32 a1=stol(attr1.substr(5));
+                nattr += bamAttrArrayWrite(a1,attr1.c_str(),attrArray+nattr);
+                break;
+            };
+            case 'A':
+            {
+                char a1=attr1.at(5);
+                nattr += bamAttrArrayWrite(a1,attr1.c_str(),attrArray+nattr);                
+                break;
+            };                
+                break;
+            case 'Z':
+            {
+                string a1=attr1.substr(5);
+                nattr += bamAttrArrayWrite(a1,attr1.c_str(),attrArray+nattr);                
+                break;
+            };
+            case 'f':
+            {
+                float a1=stof(attr1.substr(5));
+                nattr += bamAttrArrayWrite(a1,attr1.c_str(),attrArray+nattr);                
+                break;
+            };       
+        };
+    } while (pos2!= string::npos);
+    
+    return nattr;
 };
 
 template <typename intType>
@@ -215,7 +264,6 @@ int ReadAlign::alignBAM(Transcript const &trOut, uint nTrOut, uint iTrOut, uint 
             attrN+=bamAttrArrayWriteInt(trOut.nMM,"nM",attrOutArray+attrN,P);
             attrN+=bamAttrArrayWrite((to_string((uint) alignType)).at(0), "uT",attrOutArray+attrN); //cast to uint is only necessary for old compilers
             if (!P->outSAMattrRG.empty()) attrN+=bamAttrArrayWrite(P->outSAMattrRG.at(readFilesIndex),"RG",attrOutArray+attrN);
-
         } else {//this mate is mapped
             if (flagPaired) {//paired reads
                 samFLAG=0x0001;
@@ -377,6 +425,12 @@ int ReadAlign::alignBAM(Transcript const &trOut, uint nTrOut, uint iTrOut, uint 
                 };
             };
         };
+            
+        if (P->readFilesTypeN==10) {
+            if (readNameExtra[Mate].size()<1)
+                cout << iReadAll <<" " <<readName <<endl; 
+            attrN+=bamAttrArrayWriteSAMtags(readNameExtra[Mate],attrOutArray+attrN);
+        };        
 ////////////////////////////// prepare sequence and qualities
         char seqMate[DEF_readSeqLengthMax+1], qualMate[DEF_readSeqLengthMax+1];
         char *seqOut=NULL, *qualOut=NULL;